requirements.txt
aiogram
beautifulsoup4
gensim
googletrans
numpy
psycopg2_binary
Requests
google-generativeai
